Refactored map component
Pulled map initialization into a map wrapper to increase performance and readability.
Also added a store cleanup.
