+++ frontend/src/features/map_planning/hooks/useDefaultLayer.tsx
@@ -1,6 +1,9 @@
 import { useGetLayers } from './useGetLayers';
 import { LayerType } from '@/bindings/definitions';
 
+/**
+ * Gets the default layer of a specific type for the given map id.
+ */
 export function useDefaultLayer(mapId: number, layerType: LayerType) {
   const layerData = useGetLayers(mapId);
