+++ frontend/src/features/map_planning/hooks/useGetLayers.tsx
@@ -1,6 +1,10 @@
 import { getLayers } from '../api/getLayers';
 import { useQuery } from '@tanstack/react-query';
 
+/**
+ * Gets all layers for the given map id.
+ * Re-fetches data only after a certain amount of time to reduce network stress.
+ */
 export function useGetLayers(mapId: number) {
   const data = useQuery(['layers', mapId] as const, {
     queryFn: ({ queryKey: [, mapId] }) => getLayers(mapId),
