+++ frontend/src/features/map_planning/index.ts
@@ -1,1 +1,1 @@
-export * from './routes/Map';
+export * from './routes/MapWrapper';
