+++ frontend/src/features/map_planning/store/UntrackedMapStore.ts
@@ -1,38 +1,7 @@
-import type {
-  TrackedMapSlice,
-  UntrackedLayers,
-  UntrackedMapSlice,
-  UntrackedMapState,
-} from './MapStoreTypes';
-import { LayerType } from '@/bindings/definitions';
+import { TrackedMapSlice, UNTRACKED_DEFAULT_STATE, UntrackedMapSlice } from './MapStoreTypes';
 import Konva from 'konva';
 import { createRef } from 'react';
 import { StateCreator } from 'zustand';
-
-const LAYER_TYPES = Object.values(LayerType);
-
-export const UNTRACKED_DEFAULT_STATE: UntrackedMapState = {
-  mapId: -1,
-  selectedLayer: {
-    id: -1,
-    is_alternative: false,
-    name: 'none',
-    type_: LayerType.Base,
-    map_id: -1,
-  },
-  layers: {
-    ...LAYER_TYPES.reduce(
-      (acc, layerName) => ({
-        ...acc,
-        [layerName]: {
-          visible: true,
-          opacity: 1,
-        },
-      }),
-      {} as UntrackedLayers,
-    ),
-  },
-};
 
 export const createUntrackedMapSlice: StateCreator<
   TrackedMapSlice & UntrackedMapSlice,
